Cleaned up all the babel related things;
Split example into "server", "client" and "shared" example projects, to illustrate how one can mount an external Next project, while still providing a common interface for the Nest and Next apps;

Added a "useErrorHandler" module config option, which is used to specify what - if any - URLs should be rendered with Next. True for all (default), false for none, or a regex, with only matching ones being subject for error handling;

Removed RenderService.isInternalRequest() in favor of directly calling the one in the utils;

Refactored RenderService.mergeConfig() to simply use Object.assign() (and rely on TypeScript for the type safety);

The default exception filter is used for most cases where this exception filter would (re)throw an exception;

The requestHandler and errorRenderer check are moved into the constructor, to avoid repeated checks for every request.