Made the whole registration flow more aligned with NestJS's standard organization.
The module is initialized as a normal dynamic module that happens to modify global settings via ApplicationConfig. Bumped required Nest version to 6.4 to enable this. Checks for whether the Next pieces are set are removed, as they are now defined in the constructor.

Nest is now a peer dependency instead of a required one, and @nestjs/platform-fastify is only in dev dependencies, since it's an optional peer dependency.

The RenderFilter is now added instead of replacing all filters.

Fixed 404 error rendering - if the response has 404 status, the error is not serialized into Next. Instead, it is given as the symbol "Error" inside the query object. The "for" function is used, allowing applications to override this symbol with their own, enabling easier retrieval within Next.