Support specyfing index page as /
Thank you for this library. It's really lifesaving. But there's a thing which might be improved. I expected that I can use the same path for `@Get` and `@Render` decorators. But when I tried to render an index page this way `@Render('/')` I've got an error `Error: Requested and resolved page mismatch: // /`.
If I had a way to use the same paths, I would combine them in one `@Page` decorator. It would allow me to not repeat myself every time I want to render a page. The only thing required for that is to normalize given path to collapse double slashes.